update_rooms

+++ src/rooms/rooms.service.ts
@@ -1,6 +1,7 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../database/prisma.service';
 import { CreateRoomDto, UpdateRoomDto } from 'src/rooms/dto/rooms.dtos';
+import { User } from '@prisma/client';
 
 //CreatRoomDto define os dados necessarios para criar uma sala
 
@@ -8,7 +9,10 @@ import { CreateRoomDto, UpdateRoomDto } from 'src/rooms/dto/rooms.dtos';
 export class RoomsService {
   constructor(private prisma: PrismaService) {} // a propriedade prisma recebe o PrismaService permitindo a interação com o banco de dados
 
-  async create(data: CreateRoomDto) {
+  async create(data: CreateRoomDto, user: User) {
+    if (user.role !== 'ADMIN') {
+      throw new ForbiddenException ('somente admin pode criar uma sala uma sala')
+    }
     //cria a sala
     return this.prisma.room.create({ data }); //retorna os dados da sala
   }
@@ -30,7 +34,10 @@ export class RoomsService {
     return room;
   }
 
-  async update(id: string, data: UpdateRoomDto) {
+  async update(id: string, data: UpdateRoomDto, user: User) {
+    if (user.role !== 'ADMIN') {
+      throw new ForbiddenException ('somente admin pode atualizar os dados da sala')
+    }
     //atualiza os dados da sala
     return this.prisma.room.update({ where: { id }, data });
   }
@@ -44,7 +51,10 @@ export class RoomsService {
     });
   }
 
-  async remove(id: string) {
+  async remove(id: string , user:User ) {
+    if (user.role !== 'ADMIN') {
+      throw new ForbiddenException ('somente admin pode deletar uma sala uma sala')
+    }
     //remove as salas
     await this.findOne(id); // Verifica se a sala existe antes de excluir
     return this.prisma.room.delete({ where: { id } });
